+++ src/api/card/index.ts
@@ -1,0 +1,29 @@
+import { AxiosError } from 'axios';
+import axiosInstance from '..';
+import { CardDirection, CardType } from '@/components/pages/my-card/types';
+
+export async function getMyCards(
+  cardDirection: CardDirection,
+  groupIds: number[],
+  cardTypes: CardType[],
+  lastId: number,
+  size: number,
+) {
+  const urlSlug = cardDirection === 'RECEIVED' ? `received` : `sent`;
+
+  try {
+    return await axiosInstance.get(`v1/cards/${urlSlug}`, {
+      params: {
+        groupIds,
+        sid: cardTypes,
+        lastId,
+        size,
+      },
+    });
+  } catch (err: any) {
+    if (err === AxiosError) {
+      console.error(err);
+      return err.response;
+    }
+  }
+}
